% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_quantile_ensemble.R
\name{build_quantile_ensemble}
\alias{build_quantile_ensemble}
\title{Create quantile ensemble using mean or median}
\usage{
build_quantile_ensemble(
  forecast_data,
  method = c("median", "mean"),
  model_name,
  forecast_date,
  location_data = NULL
)
}
\arguments{
\item{forecast_data}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{method}{One of \code{median} (default) or \code{mean}.}

\item{model_name}{character string for \code{model} column in output}

\item{forecast_date}{a character string interpretable as a date}

\item{location_data}{am optional data.frame with metadata about location}
}
\value{
ensemble model
}
\description{
Create quantile ensemble using mean or median
}
