% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{component_outputs}
\alias{component_outputs}
\title{Example model output data for \code{linear_pool()}}
\format{
\subsection{\code{component_outputs}}{

A data frame with 123 rows and 5 columns:
\describe{
\item{model_id}{model ID}
\item{target}{forecast target}
\item{output_type}{type of forecast}
\item{output_type_id}{output type ID}
\item{value}{forecast value}
}
}
}
\usage{
component_outputs
}
\description{
Toy model output data formatted according to hubverse standards
to be used in the examples for \code{linear_pool()}. The predictions included
are taken from three normal distributions with means -3, 0, 3 and
all standard deviations 1.
}
\keyword{datasets}
