% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_model_output}
\alias{example_model_output}
\title{Example model output data}
\format{
\subsection{\code{example_model_output}}{

A data frame with 70,064 rows and 9 columns:
\describe{
\item{model_id}{unique identifier for the model submitting forecasts}
\item{location}{FIPS code identifying a location}
\item{reference_date}{the starting point of the forecasts}
\item{horizon}{number of weeks ahead being forecasted}
\item{target_end_date}{the date of occurrence of the outcome of interest;
this can be calculated directly from the \code{reference_date} and \code{horizon}
as follows: \code{target_end_date = reference_date + 7*horizon}}
\item{target}{a unique identifier for the target}
\item{output_type}{the type of representation of the prediction}
\item{output_type_id}{more identifying information specific to the output type}
\item{value}{the model’s prediction}
}
See \href{https://hubdocs.readthedocs.io/en/latest/user-guide/model-output.html}{hubverse documentation},
for more information about model output formats.
}
}
\source{
\url{https://github.com/Infectious-Disease-Modeling-Hubs/example-complex-forecast-hub/}
}
\usage{
example_model_output
}
\description{
Example model outputs adapted from forecasts submitted to the FluSight
Forecast Hub for the 2022/23 season. Model outputs data include \code{quantile},
\code{median}, \code{mean} forecasts of future incident influenza hospitalization
admissions, and \code{pmf} forecasts of hospitalization intensity,
including “low”, “moderate”, “high”, or “very high” intensities.
Hospitalization intensity categories are determined by thresholds for weekly
hospital admissions per 100,000 population. The example model output data
should be viewed only as illustrations of the data formats, not as realistic
examples of forecasts.
}
\keyword{datasets}
