% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_pool_quantile.R
\name{linear_pool_quantile}
\alias{linear_pool_quantile}
\title{Helper function for computing ensemble model outputs as a linear pool
(distributional mixture) of component model outputs for the \code{quantile}
output type.}
\usage{
linear_pool_quantile(
  model_outputs,
  weights = NULL,
  weights_col_name = "weight",
  model_id = "hub-ensemble",
  task_id_cols = NULL,
  n_samples = 10000,
  ...
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_df} with component
model outputs (e.g., predictions) with only a \code{quantile} output type.
Should be pre-validated.}

\item{weights}{an optional \code{data.frame} with component model weights. If
provided, it should have a column named \code{model_id} and a column containing
model weights. Optionally, it may contain additional columns corresponding
to task id variables, \code{output_type}, or \code{output_type_id}, if weights are
specific to values of those variables. The default is \code{NULL}, in which case
an equally-weighted ensemble is calculated. Should be pre-validated.}

\item{weights_col_name}{\code{character} string naming the column in \code{weights}
with model weights. Defaults to \code{"weight"}.}

\item{model_id}{\code{character} string with the identifier to use for the
ensemble model.}

\item{task_id_cols}{\code{character} vector with names of columns in
\code{model_outputs} that specify modeling tasks. Defaults to \code{NULL}, in which
case all columns in \code{model_outputs} other than \code{"model_id"}, the specified
\code{output_type_col} and \code{output_type_id_col}, and \code{"value"} are used as task
ids. Should be pre-validated.}

\item{n_samples}{\code{numeric} that specifies the number of samples to use when
calculating quantiles from an estimated quantile function. Defaults to \code{1e4}.
Should not be smaller than \code{1e3}.}

\item{...}{parameters that are passed to \code{distfromq::make_r_fun}, specifying
details of how to estimate a quantile function from provided quantile levels
and quantile values for \code{output_type} \code{"quantile"}.}
}
\value{
a \code{model_out_tbl} object of ensemble predictions for the \code{quantile} output type.
}
\description{
Helper function for computing ensemble model outputs as a linear pool
(distributional mixture) of component model outputs for the \code{quantile}
output type.
}
\details{
The underlying mechanism for the computations to obtain the quantiles
of a linear pool in three steps is as follows:
1. Interpolate and extrapolate from the provided quantiles for each component
model to obtain an estimate of the cdf of that distribution.
2. Draw samples from the distribution for each component model. To reduce Monte
Carlo variability, we use pseudo-random samples corresponding to quantiles
of the estimated distribution.
3. Collect the samples from all component models and extract the desired quantiles.
Steps 1 and 2 in this process are performed by \code{distfromq::make_q_fun}.
}
