% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{lop_weights}
\alias{lop_weights}
\title{Example weights data for \code{linear_pool()}}
\format{
\subsection{\code{lop_weights}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{model_id}{model ID}
\item{weight}{weight}
}
}
}
\usage{
lop_weights
}
\description{
Toy weights data formatted according to hubverse standards
to be used in the examples for \code{linear_pool()}. Weights are 0.25, 0.5, 0.25.
}
\keyword{datasets}
