% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{model_outputs}
\alias{model_outputs}
\title{Example model output data for \code{simple_ensemble()}}
\format{
\subsection{\code{model_outputs}}{

A data frame with 24 rows and 8 columns:
\describe{
\item{model_id}{model ID}
\item{location}{FIPS codes}
\item{horizon}{forecast horizon}
\item{target}{forecast target}
\item{target_date}{date that the forecast is for}
\item{output_type}{type of forecast}
\item{output_type_id}{output type ID}
\item{value}{forecast value}
}
}
}
\usage{
model_outputs
}
\description{
Toy model output data formatted according to hubverse standards
to be used in the examples for \code{simple_ensemble()}
}
\keyword{datasets}
