% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ensemble.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run ensembling methods}
\usage{
run_ensemble(
  method = "median",
  forecast_date,
  hub = "FluSight",
  data_processed_subpath = "data-forecasts/",
  criteria = c(),
  quantiles = round(c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99), 3),
  locations = covidHubUtils::hub_locations_flusight,
  target_variables = c("inc flu hosp"),
  horizons = 1:4,
  exclude_models = NULL,
  return_criteria = TRUE,
  verbose = FALSE,
  exclude_designated_other = TRUE
)
}
\arguments{
\item{method}{name of ensembling method}

\item{forecast_date}{date or character}

\item{hub}{specification of hub}

\item{data_processed_subpath}{hub-dependent, passed to load_forecasts}

\item{criteria}{filtering criteria,}

\item{quantiles}{numeric vector of quantiles}

\item{locations}{character vector of locations to ensemble}

\item{target_variables}{description of target_variables to include in ensemble}

\item{horizons}{numeric vector of horizons to include}

\item{exclude_models}{data frame of submissions such as other ensembles to be excluded}

\item{return_criteria}{return data frame with criteria passing info?}

\item{verbose}{boolean indicator of whether to print messages}

\item{exclude_designated_other}{not sure what this is.}
}
\description{
Run ensembling methods
}
\details{
Used to create a single ensemble forecast.
Takes a forecast date and loads all forecasts for the preceding week, then:
Filters models according to criteria
Ensembles forecasts according to given method
Formats ensemble forecast
Returns ensemble forecast and optionally the criteria for inclusion
}
