% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{s_ens_fweights}
\alias{s_ens_fweights}
\title{Example weights data for \code{simple_ensemble()}}
\format{
\subsection{\code{s_ens_fweights}}{

A data frame with 8 rows and 3 columns:
\describe{
\item{model_id}{model ID}
\item{location}{FIPS codes}
\item{weight}{weight}
}
}
}
\usage{
s_ens_fweights
}
\description{
Toy weights data formatted according to hubverse standards
to be used in the examples for \code{simple_ensemble()}
}
\keyword{datasets}
