% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ensemble_inputs.R
\name{validate_ensemble_inputs}
\alias{validate_ensemble_inputs}
\title{Perform simple validations on the inputs used to calculate an ensemble of
component model outputs for each combination of model task, output type,
and output type id. Valid output types should be specified by the user}
\usage{
validate_ensemble_inputs(
  model_outputs,
  weights = NULL,
  weights_col_name = "weight",
  task_id_cols = NULL,
  valid_output_types
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_df} with component
model outputs (e.g., predictions).}

\item{weights}{an optional \code{data.frame} with component model weights. If
provided, it should have a column named \code{model_id} and a column containing
model weights. Optionally, it may contain additional columns corresponding
to task id variables, \code{output_type}, or \code{output_type_id}, if weights are
specific to values of those variables. The default is \code{NULL}, in which case
an equally-weighted ensemble is calculated.}

\item{weights_col_name}{\code{character} string naming the column in \code{weights}
with model weights. Defaults to \code{"weight"}}

\item{task_id_cols}{\code{character} vector with names of columns in
\code{model_outputs} that specify modeling tasks. Defaults to \code{NULL}, in which
case all columns in \code{model_outputs} other than \code{"model_id"}, the specified
\code{output_type_col} and \code{output_type_id_col}, and \code{"value"} are used as task
ids.}

\item{valid_output_types}{\code{character} vector with the names of valid output
types for the particular ensembling method used. See the details for more
information.}
}
\value{
a list of validated model inputs: \code{model_outputs} object of class
\code{model_output_df}, optional \code{weights} data frame, \code{task_id_cols} character
vector, and \code{unique_output_types} character vector
}
\description{
Perform simple validations on the inputs used to calculate an ensemble of
component model outputs for each combination of model task, output type,
and output type id. Valid output types should be specified by the user
}
\details{
If the ensembling function intended to be used is \code{"simple_ensemble"},
the valid output types are \code{mean}, \code{median}, \code{quantile}, \code{cdf}, and \code{pmf}.
If the ensembling function will be \code{"linear_pool"}, the valid output types
are \code{mean}, \code{quantile}, \code{cdf}, \code{pmf}, and \code{sample}.
}
