% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output_type_ids.R
\name{validate_output_type_ids}
\alias{validate_output_type_ids}
\title{Perform validations to check that within each group defined by a combination
of values for task id variables and output type, all models provided the same
set of output type ids. This check only applies to the \code{cdf}, \code{pmf}, and
\code{quantile} output types to ensure the resulting distribution is valid.}
\usage{
validate_output_type_ids(model_outputs, task_id_cols)
}
\arguments{
\item{model_outputs}{an object of class \code{model_out_tbl} with component
model outputs (e.g., predictions).}

\item{task_id_cols}{\code{character} vector with names of columns in
\code{model_outputs} that specify modeling tasks.}
}
\value{
no return value
}
\description{
Perform validations to check that within each group defined by a combination
of values for task id variables and output type, all models provided the same
set of output type ids. This check only applies to the \code{cdf}, \code{pmf}, and
\code{quantile} output types to ensure the resulting distribution is valid.
}
\details{
If the ensembling function intended to be used is \code{"simple_ensemble"},
the valid output types are \code{mean}, \code{median}, \code{quantile}, \code{cdf}, and \code{pmf}.
If the ensembling function will be \code{"linear_pool"}, the valid output types
are \code{mean}, \code{quantile}, \code{cdf}, \code{pmf}, and \code{sample}.
}
