% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{weights}
\alias{weights}
\title{Example weights data for \code{linear_pool()}}
\format{
\subsection{\code{weights}}{

A data frame with 3 rows and 2 columns:
\describe{
\item{model_id}{model ID}
\item{location}{FIPS codes}
\item{weight}{weight}
}
}
}
\usage{
weights
}
\description{
Toy weights data formatted according to hubverse standards
to be used in the examples for \code{linear_pool()}
}
\keyword{datasets}
