% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_quantile_ensemble.R
\name{build_quantile_ensemble}
\alias{build_quantile_ensemble}
\title{Create quantile (weighted or unweighted) ensemble using mean or median}
\usage{
build_quantile_ensemble(
  forecast_data,
  weights_df = NULL,
  method = c("median", "mean"),
  model_name,
  forecast_date,
  location_data = NULL,
  location_col_name = "fips"
)
}
\arguments{
\item{forecast_data}{\code{data.frame} containing all the forecasts to be summarised
as an ensemble.}

\item{weights_df}{\code{data.frame} containing weights for all the models included in \code{forecast_data}
to create a weighted ensemble, should include a \code{model} column and \code{weight} column.
Defaults to NULL, which creates an unweighted ensemble.}

\item{method}{One of \code{median} (default) or \code{mean}.}

\item{model_name}{character string for \code{model} column in output; if NULL no such column will be added}

\item{forecast_date}{character string interpretable as a date which will populate the forecast_date column of
the ensemble data frame}

\item{location_data}{an optional data.frame with metadata about location. If present it will be left joined
with the ensemble data frame on the column \code{location_col_name}. This is useful when passing the ensemble to
\code{plot_forecasts} which requires full location names.}

\item{location_col_name}{character string giving the column name in \code{location_data} matching
with the \code{location} column in \code{forecast_data}. Defaults to "fips".}
}
\value{
ensemble model
}
\description{
Create quantile (weighted or unweighted) ensemble using mean or median
}
