% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_pool.R
\name{linear_pool}
\alias{linear_pool}
\title{Compute ensemble model outputs as a linear pool, otherwise known as a
distributional mixture, of component model outputs for
each combination of model task, output type, and output type id. Supported
output types include \code{mean}, \code{quantile}, \code{cdf}, \code{pmf}, and \code{sample}.}
\usage{
linear_pool(
  model_outputs,
  weights = NULL,
  weights_col_name = "weight",
  model_id = "hub-ensemble",
  task_id_cols = NULL
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_df} with component
model outputs (e.g., predictions).}

\item{weights}{an optional \code{data.frame} with component model weights. If
provided, it should have a column named \code{model_id} and a column containing
model weights. Optionally, it may contain additional columns corresponding
to task id variables, \code{output_type}, or \code{output_type_id}, if weights are
specific to values of those variables. The default is \code{NULL}, in which case
an equally-weighted ensemble is calculated.}

\item{weights_col_name}{\code{character} string naming the column in \code{weights}
with model weights. Defaults to \code{"weight"}}

\item{model_id}{\code{character} string with the identifier to use for the
ensemble model.}

\item{task_id_cols}{\code{character} vector with names of columns in
\code{model_outputs} that specify modeling tasks. Defaults to \code{NULL}, in which
case all columns in \code{model_outputs} other than \code{"model_id"}, the specified
\code{output_type_col} and \code{output_type_id_col}, and \code{"value"} are used as task
ids.}

\item{...}{parameters that are passed to \code{distfromq::make_r_fun}, specifying
details of how to estimate a quantile function from provided quantile levels and
.  quantile values.}
}
\value{
a \code{model_out_tbl} object of ensemble predictions. Note that
any additional columns in the input \code{model_outputs} are dropped.
}
\description{
Compute ensemble model outputs as a linear pool, otherwise known as a
distributional mixture, of component model outputs for
each combination of model task, output type, and output type id. Supported
output types include \code{mean}, \code{quantile}, \code{cdf}, \code{pmf}, and \code{sample}.
}
