% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_pool_sample.R
\name{make_sample_indices_unique}
\alias{make_sample_indices_unique}
\title{Make the output type ID values of sample forecasts unique for the same
combination of task IDs but different models}
\usage{
make_sample_indices_unique(model_out_tbl, numeric_output_type_ids = FALSE)
}
\arguments{
\item{model_out_tbl}{an object of class \code{model_out_tbl} with component
model outputs (e.g., predictions).}

\item{numeric_output_type_ids}{\code{logical} specifying whether the output type ID
columns should be numeric. See details for how this changes the generated
output type ID values. Defaults to FALSE.}
}
\value{
a model_out_tbl object with unique output type ID values for different
models but otherwise identical to the input model_out_tbl.
}
\description{
Make the output type ID values of sample forecasts unique for the same
combination of task IDs but different models
}
\details{
The new output type ID values will follow one of two patterns:
\enumerate{
\item If the output type ID is not numeric (may be a character):
A concatenation of the prediction's model ID and original output type ID
\item If the output type ID is numeric: A numeric representation of the above
pattern rendered as a factor.
}
}
