% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ensemble.R
\name{simple_ensemble}
\alias{simple_ensemble}
\title{Compute ensemble model outputs by summarizing component model outputs for
each combination of model task, output type, and output type id. Supported
output types include \code{mean}, \code{median}, \code{quantile}, \code{cdf}, and \code{pmf}.}
\usage{
simple_ensemble(
  model_outputs,
  weights = NULL,
  agg_fun = "mean",
  agg_args = list(),
  model_id = "hub-ensemble",
  task_id_cols = NULL,
  output_type_col = "output_type",
  output_type_id_col = "output_type_id",
  hub_connection = NULL
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_df} with component
model outputs (e.g., predictions).}

\item{weights}{an optional \code{data.frame} with component model weights. If
provided, it should have columns \code{model_id}, \code{weight}, and optionally,
additional columns corresponding to task id variables, \code{output_type}, or
\code{output_type_id}, if weights are specific to values of those variables. The
default is \code{NULL}, in which case an equally-weighted ensemble is calculated.}

\item{agg_fun}{a function or character string name of a function to use for
aggregating component model outputs into the ensemble outputs. See the
details for more information.}

\item{agg_args}{a named list of any additional arguments that will be passed
to \code{agg_fun}.}

\item{model_id}{\code{character} string with the identifier to use for the
ensemble model.}

\item{task_id_cols, }{output_type_col, output_type_id_col, value_col
\code{character} vectors with the names of the columns in \code{model_outputs} for
the output's type, additional identifying information, and value of the
model output.}
}
\value{
a data.frame with columns \code{team_abbr}, \code{model_abbr}, one column for
each task id variable, \code{output_type}, \code{output_id}, and \code{value}. Note that
any additional columns in the input \code{model_outputs} are dropped.
}
\description{
Compute ensemble model outputs by summarizing component model outputs for
each combination of model task, output type, and output type id. Supported
output types include \code{mean}, \code{median}, \code{quantile}, \code{cdf}, and \code{pmf}.
}
\details{
The default for \code{agg_fun} is \code{mean}, in which case the ensemble's
output is the average of the component model outputs within each group
defined by a combination of values in the task id columns, output type, and
output type id. The provided \code{agg_fun} should have an argument \code{x} for the
vector of numeric values to summarize, and for weighted methods, an
argument \code{w} with a numeric vector of weights. For weighted methods,
\code{agg_fun = "mean"} and \code{agg_fun = "median"} are translated to use
\code{matrixStats::weightedMean} and \code{matrixStats::weightedMedian} respectively.
}
